Moved filter logic to DataTable, and removed DataTableWithFilter.
