+++ src/components/DDItem.tsx
@@ -3,7 +3,6 @@ import { useFetchedData } from '../hooks/service';
 import { GridItem, GridItemHeader, GridItemContent } from './GridItem';
 import { Skeleton } from '@material-ui/lab';
 import DataTable from '../components/dd/DataTable';
-import DataTableWithFilter from '../components/dd/DataTableWithFilter';
 import SearchInput from '../components/SearchInput';
 import DropdownPicker from '../components/DropdownPicker';
 
@@ -119,30 +118,12 @@ export function DDChart({ payload, title, props }: DDComponentProps) {
 }
 
 export function DDTable({ payload, title, props }: DDComponentProps) {
-  return (
-    <>
-      <GridItemHeader title={title}>
-        <SearchInput />
-      </GridItemHeader>
-
-      <GridItemContent>
-        <DataTable
-          rows={payload as { rowData: any[] }[]}
-          columns={[]}
-          {...props}
-        />
-      </GridItemContent>
-    </>
-  );
-}
-
-export function DDTableWithFilter({ payload, title, props }: DDComponentProps) {
   if (
     props &&
     props['filterFunction'] &&
     props['checkBoxColumnTitle'] &&
     props['checkBoxLabel']
-  )
+  ) {
     return (
       <>
         <GridItemHeader title={title}>
@@ -150,7 +131,7 @@ export function DDTableWithFilter({ payload, title, props }: DDComponentProps) {
         </GridItemHeader>
 
         <GridItemContent>
-          <DataTableWithFilter
+          <DataTable
             rows={payload as { rowData: any[] }[]}
             columns={[]}
             filterFunction={props['filterFunction']}
@@ -161,9 +142,7 @@ export function DDTableWithFilter({ payload, title, props }: DDComponentProps) {
         </GridItemContent>
       </>
     );
-  else {
-    return (
-      <DDTable payload={payload as DDPayload} title={title} props={props} />
-    );
+  } else {
+    return <div>Hei</div>;
   }
 }
