+++ src/components/dd/DataTable.tsx
@@ -12,8 +12,9 @@ import {
 } from '@material-ui/core';
 import ExpandMoreIcon from '@material-ui/icons/ExpandMore';
 import ExpandLessIcon from '@material-ui/icons/ExpandLess';
+import { HeaderCellWithCheckBox } from '../DataTableCells';
 
-export interface DataTableColumn {
+interface DataTableColumn {
   title: string;
   expandable?: boolean;
   renderCell?: (props: { data: any; rowData: any[] }) => JSX.Element;
@@ -31,12 +32,17 @@ export interface DataTableRow {
 interface DataTableProps {
   columns: DataTableColumn[];
   rows: Omit<DataTableRow, 'columns'>[];
+  filterFunction: (
+    rows: Pick<DataTableRow, 'rowData'>
+  ) => Pick<DataTableRow, 'rowData'>;
+  checkBoxColumnTitle: string;
+  checkBoxLabel: string;
 }
 
 type DataTableRowProps = DataTableRow;
 type DataTableCellProps = DataTableCell;
 
-export const useRowStyles = makeStyles({
+const useRowStyles = makeStyles({
   root: {
     '& > *': {
       borderBottom: 'unset',
@@ -68,7 +74,7 @@ export const useRowStyles = makeStyles({
   },
 });
 
-export function Row({ rowData, columns }: DataTableRowProps) {
+function Row({ rowData, columns }: DataTableRowProps) {
   const [open, setOpen] = useState(false);
   const classes = useRowStyles();
 
@@ -128,24 +134,55 @@ export const useTableStyles = makeStyles({
   },
 });
 
-export default function DataTable({ columns, rows }: DataTableProps) {
+export default function DataTable({
+  columns,
+  rows,
+  filterFunction,
+  checkBoxColumnTitle,
+  checkBoxLabel,
+}: DataTableProps) {
   const tableClasses = useTableStyles();
   const rowClasses = useRowStyles();
+
+  const [filterFreeResources, setFilterFreeResources] = useState(false);
+
+  const handleCheckBoxChange = (event: React.ChangeEvent<HTMLInputElement>) => {
+    setFilterFreeResources(event.target.checked);
+  };
+
+  let showRows = rows;
+  if (filterFreeResources) {
+    showRows = showRows.filter((row) => filterFunction(row));
+  }
 
   return (
     <TableContainer className={tableClasses.root}>
       <Table>
         <TableHead className={tableClasses.tableHead}>
           <TableRow className={rowClasses.row}>
-            {columns.map((x) => (
-              <TableCell className={rowClasses.cell} key={x.title}>
-                {x.title}
-              </TableCell>
-            ))}
+            {columns.map((x) => {
+              if (x.title === checkBoxColumnTitle) {
+                return (
+                  <TableCell className={rowClasses.cell} key={x.title}>
+                    <HeaderCellWithCheckBox
+                      columnTitle={x.title}
+                      label={checkBoxLabel}
+                      changeHandler={handleCheckBoxChange}
+                    />
+                  </TableCell>
+                );
+              } else {
+                return (
+                  <TableCell className={rowClasses.cell} key={x.title}>
+                    {x.title}
+                  </TableCell>
+                );
+              }
+            })}
           </TableRow>
         </TableHead>
         <TableBody className={tableClasses.tableBody}>
-          {rows.map((row, i) => (
+          {showRows.map((row, i) => (
             <Row key={i} {...row} columns={columns} />
           ))}
         </TableBody>
